+++ lang/en/simplemod.php
@@ -30,19 +30,19 @@
 
 defined('MOODLE_INTERNAL') || die();
 
-$string['modulename'] = 'simplemod';
-$string['modulenameplural'] = 'simplemods';
+$string['modulename'] = 'Simplemod';
+$string['modulenameplural'] = 'Simplemods';
 $string['modulename_help'] = 'Use the simplemod module for... | The simplemod module allows...';
-$string['simplemod:addinstance'] = 'Add a new simplemod';
-$string['simplemod:submit'] = 'Submit simplemod';
-$string['simplemod:view'] = 'View simplemod';
+$string['simplemod:addinstance'] = 'Add a new Simplemod';
+$string['simplemod:submit'] = 'Submit Simplemod';
+$string['simplemod:view'] = 'View Simplemod';
 $string['simplemodfieldset'] = 'Custom example fieldset';
-$string['simplemodname'] = 'simplemod name';
-$string['simplemodname_help'] = 'This is the content of the help tooltip associated with the simplemodname field. Markdown syntax is supported.';
-$string['simplemod'] = 'simplemod';
-$string['pluginadministration'] = 'simplemod administration';
+$string['simplemodname'] = 'Name';
+$string['simplemodname_help'] = 'This is the content of the help tooltip associated with the simplemodname field.  Markdown syntax is supported.';
+$string['simplemod'] = 'Simplemod';
+$string['pluginadministration'] = 'Simplemod administration';
 $string['pluginname'] = 'simplemod';
 $string['nosimplemods'] = 'No instances';
 
 // Mod form specific simplemod settings.
-$string['title'] = 'Activity Title: ';
+$string['title'] = 'Activity title';